demo/requirements.txt
fastapi==0.109.0
google-cloud-aiplatform==1.38.1
google-auth==2.26.2
itsdangerous==2.1.2
jinja2==3.1.3
langchain==0.0.354
markdown==3.5.2
types-Markdown==3.5.0.20240106
uvicorn[standard]==0.25.0
python-multipart==0.0.6
types-protobuf==4.24.0.20240129
